+++ src/components/App.js
@@ -41,7 +41,7 @@ const helper = require('../modules/helper')
 const setting = remote.require('./setting')
 const sound = require('../modules/sound')
 const treetransformer = require('../modules/treetransformer')
-const bugout = require('../modules/bugout')
+const bugout = require('../modules/multiplayer/bugout')
 
 class App extends Component {
     constructor() {
