+++ src/components/bugout/GameLobbyModal.js
@@ -3,7 +3,7 @@ const { h, Component } = require('preact')
 // 🦹🏻‍ Bundle Bloat Protector
 import Dialog from 'preact-material-components/Dialog'
 
-const { EntryMethod, IdleStatus } = require('../../modules/bugout')
+const { EntryMethod, IdleStatus } = require('../../modules/multiplayer/bugout')
 
 class GameLobbyModal extends Component {
     constructor() {
