+++ src/modules/shims/gtp.js
@@ -2,10 +2,10 @@
 
 const EventEmitter = require('events')
 const Board = require('../board')
-const ClientId = require('../clientid')
+const ClientId = require('../multiplayer/clientid')
 const RobustWebSocket = require('robust-websocket')
 const uuidv4 = require('uuid/v4')
-const { IdleStatus, EntryMethod, emitReadyState } = require('../bugout')
+const { IdleStatus, EntryMethod, emitReadyState } = require('../multiplayer/bugout')
 
 const GATEWAY_HOST_LOCAL = "ws://localhost:3012/gateway"
 const GATEWAY_HOST_REMOTE = "wss://your.host.here:443/gateway"
